chore: Move persons join test
